added security main and uninstall

+++ uninstall.sh
@@ -9,6 +9,7 @@ check_root() {
 }
 
 remove_shc() {
+	# Remove the shc directory and its contents
 	if [ -d /opt/cstk ] && [ -e /opt/cstk/shc ]; then
 		rm -rf /opt/cstk &>/dev/null
 		if [ -f /usr/local/bin/shc ]; then
@@ -17,35 +18,51 @@ remove_shc() {
 	fi
 }
 
-remove_links() {
-	rm -f /usr/local/bin/cstk &>/dev/null
-    echo "All symbolic links and directories related to CyberSecurityToolKit have been removed."
-}
-
 remove_all() {
-	rm -rf /usr/local/bin/cstk_wrapper &>/dev/null
-	find / -type d -name CyberSecurityToolKit -exec rm -rf {} &>/dev/null \;
+	# Remove symbolic links and directories related to CyberSecurityToolKit
+	rm -f /usr/local/bin/cstk_wrapper &>/dev/null
+	rm -f /usr/local/bin/cstk &>/dev/null
+	rm -rf /opt/cstk &>/dev/null
+	find / -type d -name "CyberSecurityToolKit" -exec rm -rf {} + &>/dev/null
+    echo "All symbolic links and directories related to CyberSecurityToolKit have been removed."
 	rm -- "$0" &>/dev/null
 }
 
+remove_links() {
+	# Remove symbolic links created by the script
+	if [ -L /usr/local/bin/cstk_wrapper ]; then
+		rm -f /usr/local/bin/cstk_wrapper &>/dev/null
+	fi
+	if [ -L /usr/local/bin/cstk ]; then
+		rm -f /usr/local/bin/cstk &>/dev/null
+	fi
+	echo "All symbolic links have been removed."
+}
+
 delete_script() {
-        echo -e "1 - Remove Everything \n2 - Remove Soft Links only \n"
-        read -r -n 1 -p "==> " ans
-        if [[ "$ans" -eq 1 ]]; then
-        	echo "Uninstalling CyberSecurityToolKit..."
-			sleep 3
-			remove_links
-			remove_shc
-			remove_all
-		elif [[ "$ans" -eq 2 ]]; then
-			echo "Uninstalling CyberSecurityToolKit Links..."
-			sleep 3
-			remove_links
-		else
-			echo "Incorrect option"
-			exit
-		fi
-		echo "Uninstallation completed."
+	# Prompt user for confirmation before uninstalling
+	echo -e "1 - Remove Soft Links only \n2 - Remove all \n"
+    read -r -n 1 -p "==> " ans
+    if [[ "$ans" -eq 2 ]]; then
+        echo "Uninstalling CyberSecurityToolKit..."
+		sleep 3
+		remove_links
+		remove_shc
+		remove_all
+		echo -e "\nCyberSecurityToolKit has been uninstalled successfully."
+		echo "Thank you for using CyberSecurityToolKit!"
+		exit 0
+	elif [[ "$ans" -eq 1 ]]; then
+		echo "Uninstalling CyberSecurityToolKit Links..."
+		sleep 3
+		remove_links
+		echo -e "\nCyberSecurityToolKit links have been removed successfully."
+		echo "Thank you for using CyberSecurityToolKit!"
+		exit 0
+	else
+		echo "Incorrect option"
+		exit
+	fi
 }
 
 check_root
